+++ doc/CHANGELOG.md
@@ -23,6 +23,8 @@ will be tagged, but specific releases can also be fetched via the Git
 commit ID.
 
 ### dev
+
+### 0.0.1 (2024-07-21)
 - Added: `t8setup.bash` can now be run without setting $T8_PROJDIR if the
   current or higher directory appears to have a virtualenv with t8dev
   installed.